Issue #3517 Improved iscMosaic memory utilization and error handling when unzipping input file.

Former-commit-id: 8327c71bc422c98cde0127cbf173a0b47295388c [formerly 8327c71bc422c98cde0127cbf173a0b47295388c [formerly d39b184eea17db1f47c584d48a1448df2a0d9814]] [formerly 10a50539811423ed2db37c779aeec8e731eca851] [formerly 5a08dd10ef8e03f3eb2516a4f4715b76716acbc0 [formerly 5a08dd10ef8e03f3eb2516a4f4715b76716acbc0 [formerly e4d1cfe9e59027b0a0233c9c4a7c1ff5054dd57c]] [formerly 10a50539811423ed2db37c779aeec8e731eca851 [formerly 10a50539811423ed2db37c779aeec8e731eca851 [formerly 36d33879425d32b4ff7de617d511d7a2c51b6a4e]]]]
Former-commit-id: 5a08dd10ef8e03f3eb2516a4f4715b76716acbc0 [formerly 5a08dd10ef8e03f3eb2516a4f4715b76716acbc0 [formerly e4d1cfe9e59027b0a0233c9c4a7c1ff5054dd57c]]
Former-commit-id: 763e9a06cd6072c6acb3cbf92495d978ffb61e8e [formerly 91a3476de153e60cebb720f6c0875e3e94eb1d00] [formerly db6a7547bb20d80fe975b7ade287a6dbda801562 [formerly ad0d4d27e2747db6764b529968717b97f8515f3e]]
Former-commit-id: 0d7e04460212368b433289be87081e980e55098d [formerly 3ab0b660aca5dddd264301a18a9c9908b1c521ee]
Former-commit-id: 23575925217095fbaa2e50ca05af00034fe2f094

+++ edexOsgi/com.raytheon.edex.plugin.gfe/utility/edex_static/base/gfe/isc/iscMosaic.py
@@ -89,6 +89,7 @@ from com.raytheon.uf.edex.database.cluster import ClusterTask
 #    02/04/14        17042         ryu            Check in changes for randerso.
 #    04/11/2014      17242         David Gillingham (code checked in by zhao)
 #    07/22/2014      17484         randerso       Update cluster lock time to prevent time out
+#    08/07/2014      3517          randerso       Improved memory utilization and error handling when unzipping input file.
 #
 
 BATCH_DELAY = 0.0
@@ -538,25 +539,39 @@ class IscMosaic:
 
         gzipFile = None
         unzippedFile = None
+        gzipped = True
         try:
             import gzip
             gzipFile = gzip.open(filename, 'rb')
             unzippedFile = open(filename + ".unzipped", 'w')
-            unzippedFile.write(gzipFile.read())
+            while True:
+                buffer = gzipFile.read(65536)
+                if len(buffer) == 0:
+                    break
+                unzippedFile.write(buffer)
+        except IOError as e:
+            if e.message == "Not a gzipped file":
+                gzipped = False
+            else:
+                raise
+        else:
+            # no errors, close and rename the file
             unzippedFile.close()
             gzipFile.close()
             os.rename(unzippedFile.name, gzipFile.filename)
-        except:
-            # Not gzipped
+            gzipFile = unzippedFile = None
+        finally:
+            # close the files in case of error
             if gzipFile is not None:
                 gzipFile.close()
             if unzippedFile is not None:
                 unzippedFile.close()
-                os.remove(unzippedFile.name)
+                if not gzipped:
+                    os.remove(unzippedFile.name)
 
         a = os.times()
-        cpu = a[0] + a[1]
-        stop1 = a[4]
+        cpugz = a[0] + a[1]
+        stopgz = a[4]
 
         file = NetCDF.NetCDFFile(filename, "r")
 
@@ -656,15 +671,15 @@ class IscMosaic:
             SendNotifications.send(notification)
 
         a = os.times()
-        cpugz = a[0] + a[1]
+        cpu = a[0] + a[1]
         stop = a[4]
         logger.info("Elapsed/CPU time: "
           "%-.2f / %-.2f decompress, "
           "%-.2f / %-.2f processing, "
           "%-.2f / %-.2f total",
-          stop1 - start, cpu - cpu0,
-          stop - stop1, cpugz - cpu,
-          stop - start, cpugz - cpu0)
+          stopgz - start, cpugz - cpu0,
+          stop - stopgz, cpu - cpugz,
+          stop - start, cpu - cpu0)
 
 
     def __processParm(self, parmName, vars, history, filename):
